+++ CloudShell.ResourceDefinitions/ResourceAttributeValueShapeProvider.cs
@@ -1,0 +1,6 @@
+namespace CloudShell.ResourceDefinitions;
+
+public interface IResourceAttributeValueShapeProvider
+{
+    IReadOnlyDictionary<ResourceAttributeValueShapeId, ResourceAttributeValueShapeDefinition> GetAttributeValueShapes();
+}